+++ annotations/nocedit.exe/pseudocode/include/types/structs/SPod3Header.h
@@ -12,7 +12,7 @@ typedef struct SPod3Header {
     char author[80]; // 0x5c
     char copyright[80]; // 0xac
     int file_count; // 0xfc
-    long data_start_offset; // 0x100
+    int data_start_offset; // 0x100
     int unk1; // 0x104
     int mount_priority; // 0x108
     uint total_file_size; // 0x10c
